+++ demo/imports/routes/Login.jsx
@@ -1,10 +1,11 @@
 /* eslint-disable import/extensions, import/no-extraneous-dependencies */
-import React, { PropTypes as pt } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { pure } from 'meteor/ssrwpo:ssr';
 import LoginComponent from '../components/Login';
 
 const Login = ({ location }) => <LoginComponent location={location} />;
 Login.propTypes = {
-  location: pt.object.isRequired,
+  location: PropTypes.object.isRequired,
 };
 export default pure(Login);