+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # > 3.X See [Github's project page](https://github.com/ssrwpo/ssr/projects)
 
+# 2.5.0 PropTypes as external package
+
+* react 15.5.x deprecates `PropTypes`: user NPM's `prop-types`.
+* Meteor 1.4.4.1
+
 # 2.4.0 cordova friendly
 
 * Avoid analyzing cordova's hot code push URL: `/__cordova/`.